Forget overwrite. If the user wants to mess with the video, they can.

+++ YandereSimulatorLauncher2/MainWindow.xaml.cs
@@ -92,7 +92,10 @@ namespace YandereSimulatorLauncher2
         {
             try
             {
-                File.WriteAllBytes(inFilename, inResource);
+                if (File.Exists(inFilename) == false)
+                {
+                    File.WriteAllBytes(inFilename, inResource);
+                }
             }
             catch (Exception)
             {
